@@ -2,18 +2,18 @@
 
     <PropertyGroup>
         <OutputType>Exe</OutputType>
-        <TargetFramework>net6.0</TargetFramework>
+        <TargetFramework>net7.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="McMaster.Extensions.CommandLineUtils" Version="4.0.1"/>
-        <PackageReference Include="Spectre.Console" Version="0.45.1-preview.0.33"/>
+        <PackageReference Include="McMaster.Extensions.CommandLineUtils" Version="4.0.1" />
+        <PackageReference Include="Spectre.Console" Version="0.45.1-preview.0.33" />
     </ItemGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\ScriptCaster.services\ScriptCaster.services.csproj"/>
+        <ProjectReference Include="..\ScriptCaster.services\ScriptCaster.services.csproj" />
     </ItemGroup>
 
 </Project>
