@@ -1,26 +1,28 @@
+using System.Diagnostics;
 using ScriptCaster.Services.Helpers;
 
 namespace ScriptCaster.Services;
 
-//TODO: Change it from singleton to static. 
-public class Context
+public static class Context
 {
-    public string? TemplateName { get; private set; }
-    public string? TemplatePath { get; private set; }
-    public string? TemplateVariablePath { get; private set; }
-    public string? LocalPath { get; private set; }
-    public string? GlobalVariablePath { get; private set; }
-    public int? Recursivity { get; private set; }
-    public bool Forced { get; private set; }
-    public bool Initiated { get; private set; }
+    public static string? TemplatesCollectionPath { get; private set; }
+    public static string? TemplateName { get; private set; }
+    public static string? TemplatePath => $"{TemplatesCollectionPath}/{TemplateName}";
+    public static string? TemplateVariablePath => $"{TemplatePath}/.variables.json";
+    public static string? LocalPath { get; private set; }
+    public static string? GlobalVariablePath => "{TemplatesCollectionPath}/.variables.json";
+    public static int? Recursivity { get; private set; }
+    public static bool Forced { get; private set; }
+    public static bool Initiated { get; private set; }
 
     //TODO: the default template path shall also be a config in ~/.config/ScriptCaster/config
     //TODO: Default recursivity shall also be a config
-    public void InitContext(string? templateName, string templatesCollectionPath, int recursivity, bool forced,
+    public static void InitContext(string? templateName, string templatesCollectionPath, int recursivity, bool forced,
         bool created)
     {
         if (templateName == null) return;
 
+        TemplatesCollectionPath = templatesCollectionPath;
         TemplateName = templateName;
 
         if (!Directory.Exists(templatesCollectionPath))
@@ -32,8 +34,6 @@ public class Context
             return;
         }
 
-        TemplatePath = $"{templatesCollectionPath}/{templateName}";
-
         if (!Directory.Exists(TemplatePath) && !created)
         {
             Logger.LogError($"{TemplatePath} does not exist");
@@ -44,20 +44,18 @@ public class Context
         }
 
 
-        TemplateVariablePath = $"{TemplatePath}/.variables.json";
-
         LocalPath = Directory.GetCurrentDirectory();
 
-        GlobalVariablePath = $"{templatesCollectionPath}/.variables.json";
         Recursivity = recursivity;
         Forced = forced;
 
         Initiated = true;
     }
 
-    public static void ListTemplates(string templatesCollectionPath)
+    public static void ListTemplates()
     {
-        var templateList = DirectoryHelper.GetDirectoriesName(templatesCollectionPath);
+        Debug.Assert(TemplatesCollectionPath != null, nameof(TemplatesCollectionPath) + " should never be null... But is :<");
+        var templateList = DirectoryHelper.GetDirectoriesName(TemplatesCollectionPath);
 
         Console.WriteLine($"I found {templateList.Count()} template{(templateList.Count() > 1 ? "s" : "")} :");
 
@@ -66,19 +64,6 @@ public class Context
         if (templateList.Any()) return;
 
         Console.WriteLine("   Do not forget that only folders will be considered as template.");
-        Console.WriteLine("If you want just a file, touch it in a folder");
+        Console.WriteLine("If you want to cast file, touch it in a folder");
     }
-
-    #region Singleton
-
-    private static Context? _instance;
-
-    private Context()
-    {
-        Initiated = false;
-    }
-
-    public static Context Instance => _instance ??= new Context();
-
-    #endregion
 }
